content/media/video/test/test_progress1.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: progress events</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
var v = document.getElementById('v');
var completed = false;

function do_progress(e) {
  if (completed)
    return false;

  ok(e.lengthComputable, "Check progress lengthComputable");
  ok(e.loaded >= 0 && e.loaded <= e.total, "Check progress loaded: " + e.loaded);
  ok(e.total == 285310, "Check progress total: " + e.total);
  if (e.loaded == 285310) {
    completed = true;
    SimpleTest.finish();
  }
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
<<video id='v'
       src='seek.ogg'
       onprogress='do_progress(event)'>
</video>
</body>
</html>
